move sdk generation logic to sdk generator service
