+++ src/controllers/applicationsContentController.js
@@ -39,7 +39,6 @@ const { exec } = require('child_process');
 const { promisify } = require('util');
 const execAsync = promisify(exec);
 const os = require('os');
-const archiver = require('archiver');
 const unzipper = require('unzipper');
 const js = require('@eslint/js');
 const sdkJobService = require('../services/sdkJobService');
@@ -547,7 +546,8 @@ const generateSDK = async (req, res) => {
             });
         }
 
-        // Validate AI mode configuration
+        // Validate AI mode configuration, 
+        // TODO: need to remove this mode selection for sdk generation as only ai mode is supported
         if (!sdkConfiguration || sdkConfiguration.mode !== 'ai') {
             return res.status(400).json({
                 success: false,
@@ -567,78 +567,15 @@ const generateSDK = async (req, res) => {
         const orgID = await orgIDValue(orgName);
         
         // Create job for tracking progress
-        const job = await sdkJobService.createJob(orgID, applicationId);
+        const jobPayload = {
+            selectedAPIs,
+            sdkConfiguration,
+            orgName,
+            applicationId,
+            orgId: orgID
+        };
+        const job = await sdkJobService.createJob(orgID, applicationId, jobPayload);
         const jobId = job.JOB_ID;
-
-        // Start background processing
-        setImmediate(async () => {
-            try {
-                // Get API specifications and handles
-                const apiSpecs = await apiMetadata.getAPISpecs(orgID, selectedAPIs);
-                
-                if (apiSpecs.length === 0) {
-                    await sdkJobService.markJobAsFailed(jobId, 'No API specifications found for the selected APIs');
-                    return;
-                }
-
-                const apiHandles = await getApiHandlers(orgID, selectedAPIs);
-                const baseSdkName = sdkConfiguration?.name || `${orgName}-${applicationId}-sdk`;
-
-                // Step 1: Merging API specifications
-                await sdkJobService.startMergingStep(jobId);
-                
-                let mergedSpec;
-                if (selectedAPIs.length === 1) {
-                    mergedSpec = JSON.parse(apiSpecs[0].apiSpec);
-                } else {
-                    const mergeSpecApiRequest = prepareSDKGenerationRequest(apiSpecs, apiHandles);
-                    mergedSpec = await invokeMergeSpecApi(mergeSpecApiRequest);
-                    console.log('Merged API specifications received');
-                }
-
-                // Step 2: SDK Generation
-                await sdkJobService.startSDKGenerationStep(jobId);
-                
-                const sdkResult = await generateSDKWithOpenAPIGenerator(
-                    mergedSpec,
-                    sdkConfiguration,
-                    orgName,
-                    applicationId
-                );
-
-                console.log('SDK generated successfully, saved to:', sdkResult.sdkPath);
-
-                // Step 3: Application Code Generation
-                await sdkJobService.startAppCodeGenerationStep(jobId);
-                
-                const finalZipPath = await processAIModeSDK(sdkResult, mergedSpec, sdkConfiguration, apiHandles, baseSdkName);
-
-                // Complete job with result data
-                const resultData = {
-                    selectedAPIs: selectedAPIs,
-                    apiSpecsFound: apiSpecs.length,
-                    transformedApiHandles: apiHandles || [],
-                    sdkConfiguration: {
-                        language: sdkConfiguration?.language || 'java',
-                        name: baseSdkName,
-                        version: '1.0.0',
-                        packageName: baseSdkName,
-                        mode: 'ai',
-                        description: sdkConfiguration.description
-                    },
-                    finalDownloadUrl: `/download/sdk/${path.basename(finalZipPath)}`,
-                    applicationCodeGenerated: true,
-                    mode: 'ai',
-                    downloadPath: finalZipPath
-                };
-
-                await sdkJobService.completeJob(jobId, resultData);
-                
-            } catch (error) {
-                console.error('Error in background SDK generation:', error);
-                await sdkJobService.markJobAsFailed(jobId, error.message);
-            }
-        });
 
         // Send immediate response with job ID
         res.json({
@@ -661,590 +598,6 @@ const generateSDK = async (req, res) => {
             error: error.message
         });
     }
-};
-
-// ***** Helper Functions for SDK Generation *****
-
-
-const getApiHandlers = async (orgID, selectedAPIs) => {
-    const apiHandlesArray = [];
-    
-    for (const apiId of selectedAPIs) {
-        try {
-            const apiMetadataResult = await apiMetadata.getAPIMetadata(orgID, apiId);
-            if (apiMetadataResult && apiMetadataResult.length > 0) {
-                const apiData = apiMetadataResult[0];
-                const apiHandle = apiData.API_HANDLE || apiData.dataValues?.API_HANDLE;
-                
-                if (apiHandle) {
-                    // Transform API handle: financeapi-v1.0 -> financeapi/v1.0
-                    const transformedHandle = apiHandle.replace(/-v(\d+\.\d+)$/, '/v$1');
-                    apiHandlesArray.push(transformedHandle);
-                }
-            } else {
-                console.warn(`No metadata found for API ${apiId}`);
-            }
-        } catch (error) {
-            console.error(`Error loading metadata for API ${apiId}:`, error);
-        }
-    }
-    
-    return apiHandlesArray;
-};
-
-/**
- * Prepare the request payload for the merge OpenAPI specs API
- */
-const prepareSDKGenerationRequest = (apiSpecs, apiHandlesArray) => {
-    // Collect apiSpecString array for further processing
-    const apiSpecStrings = apiSpecs.map(spec => {
-        const parsedSpec = JSON.parse(spec.apiSpec);
-        return JSON.stringify(parsedSpec).replace(/\n/g, '');
-    });
-    
-    // Combine all specs using '\n'
-    const allSpecsString = apiSpecStrings.join('\n');
-
-    const apiHandleMap = {};
-    apiHandlesArray.forEach((apiHandle, index) => {
-        apiHandleMap[`spec${index + 1}`] = apiHandle;
-    });
-    
-    return {
-        "specifications": allSpecsString,
-        "contexts": JSON.parse(JSON.stringify(apiHandleMap))
-    };
-}
-
-/**
- * Invoke the backend merge OpenAPI specs API
- */
-const invokeMergeSpecApi = async(requestPayload) => {
-    try {
-        const response = await fetch(`${aiSDKServiceUrl}${aiSDKServiceEndpoints.mergeSpecs}`, {
-            method: 'POST',
-            headers: {
-                'Content-Type': 'application/json'
-            },
-            body: JSON.stringify(requestPayload)
-        });
-        if (!response.ok) {
-            throw new Error(`Failed to merge API specifications: ${response.statusText}`);
-        }
-        const data = await response.json();
-        return data;
-    } catch (error) {
-        console.error("Error invoking merge spec API:", error);
-        throw error;
-    }
-};
-
-/**
- * Process AI mode SDK generation - extract SDK, generate application code, and create final ZIP
- */
-const processAIModeSDK = async (sdkResult, mergedSpec, sdkConfiguration, apiHandles, baseSdkName) => {
-    try {
-        console.log('Processing AI mode SDK generation...');
-        
-        // Extract SDK methods for AI processing
-        const sdkMethods = await extractMethodFile(sdkResult.sdkPath, sdkConfiguration?.language || 'java');
-        
-        // Generate application code using AI
-        const applicationApiRequest = await getApplicationGenApiReqBody(
-            sdkMethods.apiClassContent,
-            sdkConfiguration?.language || 'java',
-            mergedSpec,
-            sdkConfiguration
-        );
-        
-        const applicationCodeResponse = await invokeApplicationCodeGenApi(applicationApiRequest);
-        
-        // Create final ZIP with SDK + application code
-        const finalZipName = createSDKFileName(apiHandles, baseSdkName, true);
-        const finalZipResult = await processApplicationCodeAndCreateFinalZip(
-            applicationCodeResponse,
-            sdkResult.sdkPath,
-            sdkConfiguration,
-            finalZipName
-        );
-        
-        return finalZipResult.finalZipPath;
-        
-    } catch (error) {
-        console.error('Error processing AI mode SDK:', error);
-        
-        // Ensure cleanup of SDK folder even if processing fails
-        try {
-            await fs.promises.rm(sdkResult.sdkPath, { recursive: true, force: true });
-            console.log('Cleaned up SDK folder after AI mode error');
-        } catch (cleanupError) {
-            console.error('Error cleaning up SDK folder:', cleanupError);
-        }
-        
-        throw error;
-    }
-};
-
-/**
- * Create SDK filename based on API handles and mode
- */
-const createSDKFileName = (apiHandles, baseSdkName, isAIMode) => {
-    const apiHandlesSuffix = apiHandles.map(handle => 
-        handle.replace(/[^a-zA-Z0-9]/g, '') // Remove special characters
-    ).join('-');
-    
-    const suffix = isAIMode ? '-with-app' : '';
-    return `${baseSdkName}-${apiHandlesSuffix}${suffix}.zip`;
-};
-
-/**
- * Generate SDK using openapi-generator
- */
-const generateSDKWithOpenAPIGenerator = async (mergedSpec, sdkConfiguration, orgName, applicationId) => {
-    let tempDir = null;
-    let outputDir = null;
-    
-    try {
-        console.log('Starting SDK generation with OpenAPI Generator...');
-        const language = sdkConfiguration?.language || 'javascript';
-        const sdkName = sdkConfiguration?.name || `${orgName}-${applicationId}-sdk`;
-        
-        // Create temporary directory for SDK generation
-        tempDir = path.join(os.tmpdir(), 'sdk-generation', `${sdkName}-${Date.now()}`);
-        const specFilePath = path.join(tempDir, 'merged-spec.json');
-        outputDir = path.join(process.cwd(), 'generated-sdks', `${sdkName}-${Date.now()}`);
-        
-        // Create directories
-        await fs.promises.mkdir(tempDir, { recursive: true });
-        await fs.promises.mkdir(outputDir, { recursive: true });
-        
-        // Write merged spec to file
-        await fs.promises.writeFile(specFilePath, JSON.stringify(mergedSpec, null, 2));
-        
-        // Determine the generator name based on language
-        const generatorMap = {
-            'javascript': 'javascript',
-            'java': 'java'
-        };
-        
-        const generator = generatorMap[language] || 'javascript';
-        
-        // Build openapi-generator command
-        const command = [
-            'openapi-generator-cli',
-            'generate',
-            '-i', specFilePath,
-            '-g', generator,
-            '-o', outputDir,
-            '--package-name', sdkName,
-            '--additional-properties',
-            `packageName=${sdkName},projectName=${sdkName},packageVersion=1.0.0`
-            ].join(' ');
-        
-        console.log(`Generating SDK with command: ${command}`);
-        
-        // Execute openapi-generator
-        const { stdout, stderr } = await execAsync(command, { 
-            cwd: tempDir,
-            timeout: 120000 // 2 minutes timeout
-        });
-        
-        if (stderr && !stderr.includes('WARN')) {
-            console.warn('openapi-generator stderr:', stderr);
-        }
-                
-        // Clean up temp directory
-        //await fs.promises.rm(tempDir, { recursive: true, force: true });
-        
-        return {
-            success: true,
-            sdkPath: outputDir,
-            language: language,
-            sdkName: sdkName
-        };
-        
-    } catch (error) {
-        console.error('Error generating SDK with openapi-generator:', error);
-        
-        // Provide more specific error messages based on error type
-        let errorMessage = 'SDK generation failed';
-        
-        if (error.message && error.message.includes('NullPointerException')) {
-            errorMessage = 'SDK generation failed due to invalid OpenAPI specification. The specification may be missing required fields.';
-        } else if (error.message && error.message.includes('Command failed')) {
-            errorMessage = 'OpenAPI Generator failed to process the specification. Please check the API specification format.';
-        } else if (error.code === 'ENOENT') {
-            errorMessage = 'OpenAPI Generator CLI not found. Please install openapi-generator-cli: npm install -g @openapitools/openapi-generator-cli';
-        } else if (error.message && error.message.includes('timeout')) {
-            errorMessage = 'SDK generation timed out. The API specification might be too complex.';
-        } else if (error.message) {
-            errorMessage = error.message;
-        }
-        
-        // Clean up directories on error
-        try {
-            if (tempDir) {
-                //await fs.promises.rm(tempDir, { recursive: true, force: true });
-                console.log('Cleaned up temp directory after SDK generation error');
-            }
-            if (outputDir) {
-                //await fs.promises.rm(outputDir, { recursive: true, force: true });
-                console.log('Cleaned up output directory after SDK generation error');
-            }
-        } catch (cleanupError) {
-            console.error('Error cleaning up directories after SDK generation error:', cleanupError);
-        }
-        
-        throw new Error(errorMessage);
-    }
-};
-
-/**
- * Create ZIP archive of the generated SDK
- */
-const createZipArchive = async (sourceDir, outputPath) => {
-    return new Promise((resolve, reject) => {
-        const output = fs.createWriteStream(outputPath);
-        const archive = archiver('zip', { zlib: { level: 9 } });
-        
-        output.on('close', () => {
-            console.log(`ZIP archive created: ${archive.pointer()} total bytes`);
-            resolve();
-        });
-        
-        archive.on('error', (err) => {
-            reject(err);
-        });
-        
-        archive.pipe(output);
-        archive.directory(sourceDir, false);
-        archive.finalize();
-    });
-};
-
-/**
- * Extract ZIP and process SDK files to send to backend API
- */
-const extractMethodFile = async (sdkDir, language) => {
-    try {
-        
-        // Find the API class file based on language
-        let apiClassFile = null;
-        let apiClassContent = null;
-        
-        if (language === 'java') {
-            // Look for DefaultApi.java file in src/main/java directory structure
-            const javaApiPath = await findFileRecursively(sdkDir, 'DefaultApi.java');
-            if (javaApiPath) {
-                apiClassFile = javaApiPath;
-                apiClassContent = await fs.promises.readFile(javaApiPath, 'utf8');
-                console.log(`Found Java API class: ${javaApiPath}`);
-            }
-        } else if (language === 'javascript' || language === 'android') {
-            // Look for API class in JavaScript/TypeScript SDKs
-            const jsApiPath = await findFileRecursively(sdkDir, /.*[Aa]pi\.js$|.*[Aa]pi\.ts$/);
-            if (jsApiPath) {
-                apiClassFile = jsApiPath;
-                apiClassContent = await fs.promises.readFile(jsApiPath, 'utf8');
-                console.log(`Found JS/TS API class: ${jsApiPath}`);
-            }
-        } else if (language === 'python') {
-            // Look for API class in Python SDKs
-            const pythonApiPath = await findFileRecursively(sdkDir, /.*_api\.py$|.*api\.py$/);
-            if (pythonApiPath) {
-                apiClassFile = pythonApiPath;
-                apiClassContent = await fs.promises.readFile(pythonApiPath, 'utf8');
-                console.log(`Found Python API class: ${pythonApiPath}`);
-            }
-        }
-        
-        if (!apiClassContent) {
-            console.warn(`No API class file found for language: ${language}`);
-        }
-        
-        return {
-            apiClassContent: apiClassContent || '',
-            apiClassFile: path.basename(apiClassFile || 'UnknownApi'),
-            apiClassFound: !!apiClassContent,
-            SdkPath: sdkDir
-        };
-        
-    } catch (error) {
-        console.error('Error extracting and processing SDK:', error);
-        throw error;
-    }
-};
-
-/**
- * Recursively find a file by name or pattern
- */
-const findFileRecursively = async (dir, filePattern) => {
-    try {
-        const items = await fs.promises.readdir(dir, { withFileTypes: true });
-        
-        for (const item of items) {
-            const fullPath = path.join(dir, item.name);
-            
-            if (item.isDirectory()) {
-                const result = await findFileRecursively(fullPath, filePattern);
-                if (result) return result;
-            } else if (item.isFile()) {
-                if (typeof filePattern === 'string') {
-                    if (item.name === filePattern) {
-                        return fullPath;
-                    }
-                } else if (filePattern instanceof RegExp) {
-                    if (filePattern.test(item.name)) {
-                        return fullPath;
-                    }
-                }
-            }
-        }
-        
-        return null;
-    } catch (error) {
-        console.error(`Error searching in directory ${dir}:`, error);
-        return null;
-    }
-};
-
-/**
- * Find the Java package directory where API classes are located
- * This looks for the directory that contains api/, auth/, and model/ subdirectories
- */
-const findJavaPackageDirectory = async (sdkDir) => {
-    try {
-        console.log(`Searching for Java package directory in: ${sdkDir}`);
-        
-        // Common Java package paths in OpenAPI generated SDKs
-        const possiblePaths = [
-            'src/main/java/org/openapitools/client',
-            'src/main/java/io/swagger/client',
-            'src/main/java/com/example/client',
-            'src/main/java'
-        ];
-        
-        // First, try the common OpenAPI tools structure
-        for (const possiblePath of possiblePaths) {
-            const fullPath = path.join(sdkDir, possiblePath);
-            console.log(`Checking path: ${fullPath}`);
-            
-            if (await fs.promises.access(fullPath).then(() => true).catch(() => false)) {
-                console.log(`Path exists: ${fullPath}`);
-                
-                // Check if this directory contains the expected subdirectories
-                const hasApiDir = await fs.promises.access(path.join(fullPath, 'api')).then(() => true).catch(() => false);
-                const hasAuthDir = await fs.promises.access(path.join(fullPath, 'auth')).then(() => true).catch(() => false);
-                const hasModelDir = await fs.promises.access(path.join(fullPath, 'model')).then(() => true).catch(() => false);
-                
-                console.log(`Directory contents check - api: ${hasApiDir}, auth: ${hasAuthDir}, model: ${hasModelDir}`);
-                
-                if (hasApiDir || hasAuthDir || hasModelDir) {
-                    console.log(`Found Java package directory: ${fullPath}`);
-                    return fullPath;
-                }
-            } else {
-                console.log(`Path does not exist: ${fullPath}`);
-            }
-        }
-        
-        // If common paths don't work, search recursively for a directory that contains api/auth/model subdirs
-        console.log('Searching recursively for Java package structure...');
-        console.log('Current SDK directory structure:');
-        await listDirectoryStructure(sdkDir);
-        
-        const foundPath = await findJavaPackageRecursively(sdkDir);
-        if (foundPath) {
-            console.log(`Found Java package directory recursively: ${foundPath}`);
-            return foundPath;
-        }
-        
-        console.warn('Could not find Java package directory structure');
-        return null;
-        
-    } catch (error) {
-        console.error('Error finding Java package directory:', error);
-        return null;
-    }
-};
-
-/**
- * Recursively search for Java package directory
- */
-const findJavaPackageRecursively = async (dir, maxDepth = 10, currentDepth = 0) => {
-    if (currentDepth > maxDepth) {
-        return null;
-    }
-    
-    try {
-        const items = await fs.promises.readdir(dir, { withFileTypes: true });
-        
-        // Check if current directory has api, auth, or model subdirectories
-        const subdirs = items.filter(item => item.isDirectory()).map(item => item.name);
-        const hasApiDir = subdirs.includes('api');
-        const hasAuthDir = subdirs.includes('auth');
-        const hasModelDir = subdirs.includes('model');
-        
-        // If we found at least one of the expected directories, this is likely the package root
-        if (hasApiDir || hasAuthDir || hasModelDir) {
-            return dir;
-        }
-        
-        // Continue searching in subdirectories
-        for (const item of items) {
-            if (item.isDirectory() && item.name !== 'api' && item.name !== 'auth' && item.name !== 'model') {
-                const fullPath = path.join(dir, item.name);
-                const result = await findJavaPackageRecursively(fullPath, maxDepth, currentDepth + 1);
-                if (result) {
-                    return result;
-                }
-            }
-        }
-        
-        return null;
-    } catch (error) {
-        console.error(`Error searching directory ${dir}:`, error);
-        return null;
-    }
-};
-
-/**
- * Create JSON request data for backend API request
- */
-const getApplicationGenApiReqBody = async (apiClassContent, language, mergedSpec, sdkConfiguration) => {
-    let useCase;
-    
-    if (sdkConfiguration?.mode === 'ai' && sdkConfiguration?.description && sdkConfiguration.description.trim()) {
-        useCase = sdkConfiguration.description.trim();
-        console.log('Using AI mode with user prompt:', useCase);
-    } else {
-        // Default use case for default mode
-        useCase = `Generate application which uses the provided APIs to demonstrate their functionality with sample data and proper error handling`;
-        console.log('Using default mode with standard use case');
-    }
-    
-    return {
-        sdkMethodsFile: apiClassContent || '',
-        useCase: useCase,
-        language: language,
-        APISpecification: JSON.stringify(mergedSpec)
-    };
-};
-
-const invokeApplicationCodeGenApi = async (requestData) => {
-    try {
-        const response = await fetch(`${aiSDKServiceUrl}${aiSDKServiceEndpoints.generateApp}`, {
-            method: 'POST',
-            headers: {
-                'Content-Type': 'application/json'
-            },
-            body: JSON.stringify(requestData)
-        });
-        if (!response.ok) {
-            throw new Error(`Failed to generate application: ${response.statusText}`);
-        }
-        const data = await response.text();
-        return data;
-    } catch (error) {
-        console.error("Error invoking application generation API:", error);
-        throw error;
-    }
-};
-
-/**
- * Process application code response and create final ZIP with SDK + application code
- */
-const processApplicationCodeAndCreateFinalZip = async (applicationCode, sdkPath, sdkConfiguration, zipFileName) => {
-    try {
-        console.log('Processing application code response and creating final ZIP...');
-        const language = sdkConfiguration?.language || 'java';
-        
-        // Create application code files in the SDK folder
-        if (applicationCode) {
-            const fileExtension = getFileExtension(language);
-            
-            if (language === 'java') {
-                // For Java projects, place Application.java in the correct package structure
-                console.log('Searching for Java package directory structure...');
-                const javaPackageDir = await findJavaPackageDirectory(sdkPath);
-                
-                if (javaPackageDir) {
-                    // Place Application.java at the same level as api, auth, model packages
-                    const mainAppFile = path.join(javaPackageDir, `Application.${fileExtension}`);
-                    await fs.promises.writeFile(mainAppFile, applicationCode, 'utf8');
-                    console.log(`Java Application code written to: ${mainAppFile}`);
-                    
-                    // Also log the relative path for clarity
-                    const relativePath = path.relative(sdkPath, mainAppFile);
-                    console.log(`Relative path: ${relativePath}`);
-                } else {
-                    // Fallback: create in application-code directory
-                    console.warn('Could not find Java package directory, using fallback location');
-                    const appCodeDir = path.join(sdkPath, 'application-code');
-                    await fs.promises.mkdir(appCodeDir, { recursive: true });
-                    const mainAppFile = path.join(appCodeDir, `Application.${fileExtension}`);
-                    await fs.promises.writeFile(mainAppFile, applicationCode, 'utf8');
-                    console.log(`Application code written to fallback location: ${mainAppFile}`);
-                }
-            } else {
-                // For non-Java languages, use the application-code directory
-                const appCodeDir = path.join(sdkPath, 'application-code');
-                await fs.promises.mkdir(appCodeDir, { recursive: true });
-                const mainAppFile = path.join(appCodeDir, `Application.${fileExtension}`);
-                await fs.promises.writeFile(mainAppFile, applicationCode, 'utf8');
-                console.log(`Application code written to: ${mainAppFile}`);
-            }
-        }
-        
-        // Create README with instructions based on language and structure
-        await createProjectReadme(sdkPath, sdkConfiguration, language);
-        
-        // Create final ZIP with SDK + application code
-        const permanentDir = path.join(process.cwd(), 'generated-sdks');
-        await fs.promises.mkdir(permanentDir, { recursive: true });
-        
-        const finalZipPath = path.join(permanentDir, zipFileName);
-        await createZipArchive(sdkPath, finalZipPath);
-        console.log(`Final ZIP created: ${finalZipPath}`);
-        
-        // Clean up SDK folder
-        await fs.promises.rm(sdkPath, { recursive: true, force: true });
-        console.log('Cleaned up temporary SDK folder');
-        
-        return {
-            finalZipPath: finalZipPath,
-        };
-        
-    } catch (error) {
-        console.error('Error processing application code and creating final ZIP:', error);
-        
-        // Ensure cleanup of SDK folder even if ZIP creation fails
-        try {
-            await fs.promises.rm(sdkPath, { recursive: true, force: true });
-            console.log('Cleaned up SDK folder after ZIP creation error');
-        } catch (cleanupError) {
-            console.error('Error cleaning up SDK folder:', cleanupError);
-        }
-        
-        throw error;
-    }
-};
-
-/**
- * Get file extension based on programming language
- */
-const getFileExtension = (language) => {
-    const extensionMap = {
-        'java': 'java',
-        'javascript': 'js',
-        'typescript': 'ts',
-        'python': 'py',
-        'csharp': 'cs',
-        'go': 'go',
-        'php': 'php',
-        'ruby': 'rb'
-    };
-    return extensionMap[language] || 'java';
 };
 
 /**
@@ -1300,175 +653,6 @@ const cleanupOrphanedSDKDirectories = async () => {
         
     } catch (error) {
         console.error('Error during orphaned SDK directories cleanup:', error);
-    }
-};
-
-/**
- * Create a comprehensive README file for the generated SDK and application
- */
-const createProjectReadme = async (sdkPath, sdkConfiguration, language) => {
-    try {
-        const sdkName = sdkConfiguration?.name || 'generated-sdk';
-        const mode = sdkConfiguration?.mode || 'default';
-        
-        let readmeContent = `# ${sdkName}
-
-Generated SDK with Application Code
-
-## Overview
-This package contains:
-1. **SDK Library**: Generated from your selected API specifications
-2. **Application Code**: Sample application demonstrating API usage
-3. **Documentation**: This README and any additional documentation
-
-## Configuration
-- **Language**: ${language}
-- **SDK Name**: ${sdkName}
-- **Generation Mode**: ${mode}
-- **Generated**: ${new Date().toISOString()}
-
-`;
-
-        if (language === 'java') {
-            readmeContent += `## Java Project Structure
-
-The generated Java SDK follows standard Maven/Gradle project structure:
-
-\`\`\`
-src/
-├── main/
-│   ├── AndroidManifest.xml (for Android projects)
-│   └── java/
-│       └── org/openapitools/client/
-│           ├── Application.java          # ← Your generated application
-│           ├── ApiClient.java           # Main API client
-│           ├── Configuration.java       # SDK configuration
-│           ├── api/                     # API endpoints
-│           ├── auth/                    # Authentication classes
-│           └── model/                   # Data models
-├── build.gradle (or pom.xml)
-└── README.md
-\`\`\`
-
-## Quick Start (Java)
-
-### 1. Build the Project
-\`\`\`bash
-# For Gradle projects
-./gradlew build
-
-# For Maven projects
-mvn clean compile
-\`\`\`
-
-### 2. Run the Application
-\`\`\`bash
-# Compile and run Application.java
-javac -cp "lib/*:src/main/java" src/main/java/org/openapitools/client/Application.java
-java -cp "lib/*:src/main/java" org.openapitools.client.Application
-
-# Or use your build tool
-./gradlew run
-# or
-mvn exec:java -Dexec.mainClass="org.openapitools.client.Application"
-\`\`\`
-
-### 3. Integration
-To use this SDK in your own project:
-
-1. **Add the SDK as a dependency** in your \`build.gradle\` or \`pom.xml\`
-2. **Import the necessary classes**:
-   \`\`\`java
-   import org.openapitools.client.ApiClient;
-   import org.openapitools.client.Configuration;
-   import org.openapitools.client.api.*;
-   import org.openapitools.client.model.*;
-   \`\`\`
-3. **Initialize the API client**:
-   \`\`\`java
-   ApiClient client = Configuration.getDefaultApiClient();
-   client.setBasePath("https://your-api-server.com");
-   \`\`\`
-
-## Application.java
-The generated \`Application.java\` file contains:
-- Complete working examples of API calls
-- Error handling patterns
-- Authentication setup (if required)
-- Sample data processing
-
-`;
-        } else {
-            readmeContent += `## ${language.charAt(0).toUpperCase() + language.slice(1)} Project
-
-Please refer to the application code in the application-code directory.
-
-`;
-        }
-
-        readmeContent += `## API Documentation
-- Check the docs/ directory for detailed API documentation
-- Each API endpoint is documented with request/response examples
-- Model classes are documented with field descriptions
-
-## Authentication
-If your APIs require authentication, update the configuration in Application.${getFileExtension(language)}:
-- API Keys: Set in the ApiClient configuration
-- OAuth2: Configure the OAuth2 flow
-- Basic Auth: Set username/password in the client
-
-## Troubleshooting
-
-### Common Issues
-1. **Compilation Errors**: Ensure all dependencies are installed
-2. **Runtime Errors**: Check API endpoint URLs and authentication
-3. **Network Issues**: Verify server connectivity and firewall settings
-
-### Getting Help
-- Check the generated documentation in docs/ folder
-- Review the API specification used for generation
-- Examine the sample code in Application.${getFileExtension(language)}
-
-## License
-This generated code is provided as-is for demonstration purposes.
-Check your API provider's terms of service for usage guidelines.
-`;
-
-        const readmePath = path.join(sdkPath, 'README.md');
-        await fs.promises.writeFile(readmePath, readmeContent, 'utf8');
-        console.log(`Comprehensive README created: ${readmePath}`);
-        
-    } catch (error) {
-        console.error('Error creating project README:', error);
-        // Don't throw error, as this is not critical for SDK generation
-    }
-};
-
-/**
- * Debug function to list directory structure (helpful for troubleshooting)
- */
-const listDirectoryStructure = async (dir, prefix = '', maxDepth = 3, currentDepth = 0) => {
-    if (currentDepth > maxDepth) {
-        return;
-    }
-    
-    try {
-        const items = await fs.promises.readdir(dir, { withFileTypes: true });
-        
-        for (const item of items.slice(0, 20)) { // Limit to first 20 items to avoid spam
-            console.log(`${prefix}${item.isDirectory() ? '📁' : '📄'} ${item.name}`);
-            
-            if (item.isDirectory() && currentDepth < maxDepth) {
-                const fullPath = path.join(dir, item.name);
-                await listDirectoryStructure(fullPath, prefix + '  ', maxDepth, currentDepth + 1);
-            }
-        }
-        
-        if (items.length > 20) {
-            console.log(`${prefix}... and ${items.length - 20} more items`);
-        }
-    } catch (error) {
-        console.log(`${prefix} Error reading directory: ${error.message}`);
     }
 };
